mcnode: refactor net/directory specifics into mcnode/net
